Added configuration for setting xbee data rate

+++ src/lib/comms/xbee/xbee.h
@@ -67,7 +67,7 @@ class XBeeDevice
     void startup(const std::string& port_name, const int baud_rate, const NodeId& my_node_id,
                  const uint16_t network_id, const std::string& xbee_info_location,
                  const bool& use_encryption, const std::string& encryption_password,
-                 const uint8_t rf_datarate);
+                 const uint16_t rf_datarate);
     void shutdown();
 
     std::vector<NodeId> get_peers();
